requirements.txt
Django==3.2.9
gunicorn
django-dotenv==1.4.2
django-cors-headers==3.10.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
Pillow==8.4.0
python3-openid==3.2.0
six==1.16.0